# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/services/app/libitem/ImportExcel": {
            "post": {
                "description": "通过提供的Excel文件导入图书信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "libitems"
                ],
                "summary": "导入图书信息",
                "parameters": [
                    {
                        "description": "导入参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ImportExcelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "导入成功，返回导入的数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ImportExcelDto"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的请求",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/services/app/order/CreateOrder": {
            "post": {
                "description": "创建新的订单并返回订单信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "订单信息",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TestOrder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功创建的订单信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TestOrder"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的请求",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/services/app/order/GetOrders": {
            "get": {
                "description": "获取所有订单信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "获取所有订单",
                "responses": {
                    "200": {
                        "description": "订单列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.TestOrder"
                                            }
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/services/app/user/CreateUser": {
            "post": {
                "description": "创建一个新的用户并返回用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "创建新用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功创建的用户信息",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserDto"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的输入",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "创建用户失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/services/app/user/GetUsers": {
            "get": {
                "description": "获取所有用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "获取所有用户",
                "responses": {
                    "200": {
                        "description": "获取到的所有用户列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.UserDto"
                                            }
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.TestOrder": {
            "description": "订单数据模型",
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 100.5
                },
                "customer": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "description": "@ID",
                    "type": "integer"
                }
            }
        },
        "request.ImportExcelInput": {
            "type": "object",
            "required": [
                "path",
                "tenantid"
            ],
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string",
                    "default": "作者"
                },
                "barcode": {
                    "description": "条码号",
                    "type": "string",
                    "default": "条码号"
                },
                "callNo": {
                    "description": "索书号",
                    "type": "string",
                    "default": "索书号"
                },
                "catalogCode": {
                    "description": "分类号",
                    "type": "string",
                    "default": "分类号"
                },
                "isbn": {
                    "description": "ISBN",
                    "type": "string",
                    "default": "ISBN"
                },
                "locationname": {
                    "description": "当前馆藏地点",
                    "type": "string",
                    "default": "当前馆藏地点"
                },
                "pages": {
                    "description": "页数",
                    "type": "string",
                    "default": "页数"
                },
                "path": {
                    "description": "excel 服务器本地路径",
                    "type": "string",
                    "default": "d:/4.xlsx",
                    "maxLength": 100,
                    "minLength": 3
                },
                "price": {
                    "description": "单价",
                    "type": "string",
                    "default": "单价"
                },
                "pubDate": {
                    "description": "出版日期",
                    "type": "string",
                    "default": "出版日期"
                },
                "publisher": {
                    "description": "出版社",
                    "type": "string",
                    "default": "出版社"
                },
                "tenantid": {
                    "description": "租户",
                    "type": "integer",
                    "default": 1,
                    "maximum": 99999,
                    "minimum": 1
                },
                "tid": {
                    "description": "tid",
                    "type": "string",
                    "default": "tid"
                },
                "title": {
                    "description": "题名",
                    "type": "string",
                    "default": "题名"
                }
            }
        },
        "request.UserRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "response.ImportExcelDto": {
            "type": "object"
        },
        "response.UserDto": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                },
                "name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}